Refactoring to split service implementation to a runner
